backend/requirements.txt
fastapi
uvicorn
anthropic
python-multipart
Pillow
sqlalchemy
pyjwt
